[LibOS] Support special keys (MRENCLAVE, MRSIGNER) in `encrypted` fs
Key names beginning with `_` are considered special, and PAL is queried
for these keys. This is used for SGX MRENCLAVE and MRSIGNER keys.

If a key is not supported by the current PAL, mounting will still
succeed (but files will not be readable). This way, manifests using
these PAL-host-specific keys can still be loaded when using other PAL
hosts.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>